+++ client/src/GalleryPage/mapObjects/Light.tsx
